gedeeld_met_eenheden en gedeeld_met_onzelfstandige_woonruimten in functies gezet
