+++ woningwaardering/stelsels/zelfstandige_woonruimten/oppervlakte_van_overige_ruimten/oppervlakte_van_overige_ruimten.py
@@ -6,9 +6,12 @@ from typing import Iterator
 from loguru import logger
 
 from woningwaardering.stelsels import Stelselgroep
-from woningwaardering.stelsels.utils import naar_tabel, rond_af, rond_af_op_kwart
-from woningwaardering.stelsels.zelfstandige_woonruimten.utils import (
+from woningwaardering.stelsels.utils import (
     classificeer_ruimte,
+    gedeeld_met_eenheden,
+    naar_tabel,
+    rond_af,
+    rond_af_op_kwart,
     voeg_oppervlakte_kasten_toe_aan_ruimte,
 )
 from woningwaardering.vera.bvg.generated import (
@@ -65,8 +68,7 @@ class OppervlakteVanOverigeRuimten(Stelselgroep):
         ruimten = [
             ruimte
             for ruimte in eenheid.ruimten or []
-            if ruimte.gedeeld_met_aantal_eenheden is None
-            or ruimte.gedeeld_met_aantal_eenheden == 1
+            if not gedeeld_met_eenheden(ruimte)
         ]
 
         for ruimte in ruimten:
