+++ woningwaardering/stelsels/zelfstandige_woonruimten/sanitair/sanitair.py
@@ -60,8 +60,7 @@ class Sanitair(Stelselgroep):
         ruimten = [
             ruimte
             for ruimte in eenheid.ruimten or []
-            if ruimte.gedeeld_met_aantal_eenheden is None
-            or ruimte.gedeeld_met_aantal_eenheden < 2
+            if not utils.gedeeld_met_eenheden(ruimte)
         ]
 
         for ruimte in ruimten:
