+++ woningwaardering/stelsels/zelfstandige_woonruimten/verkoeling_en_verwarming/verkoeling_en_verwarming.py
@@ -11,7 +11,7 @@ from woningwaardering.stelsels.criteriumsleutels import (
     CriteriumSleutels,
 )
 from woningwaardering.stelsels.stelselgroep import Stelselgroep
-from woningwaardering.stelsels.zelfstandige_woonruimten.utils import classificeer_ruimte
+from woningwaardering.stelsels.utils import classificeer_ruimte
 from woningwaardering.vera.bvg.generated import (
     EenhedenEenheid,
     EenhedenRuimte,
